[MOSIP-43633] Code changes for removing the cyclic dependency from the biometric update bug fix code. (#2221)
* [MOSIP-43633] Code changes for removing the cyclic dependency from the biometric update bug fix code.


* [MOSIP-43633] - UtilitiesTest to UtilityTest movement


---------
