+++ registration-processor/core-processor/registration-processor-bio-dedupe-stage/src/test/java/io/mosip/registration/processor/biodedupe/stage/BioDedupeProcessorTest.java
@@ -738,9 +738,9 @@ public class BioDedupeProcessorTest {
 
 		when(abisHandlerUtil.getProcessedMatchedResult(anyString(), anyString(), anyInt(), anyString(), any()))
 				.thenReturn(processedMatchedResult);
-		when(utilities.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
+		when(utility.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
 				.thenReturn(false);
-		when(utilities.allBiometricHaveException(anyString(), anyString(), any()))
+		when(utility.allBiometricHaveException(anyString(), anyString(), any()))
 				.thenReturn(false);
 
 		MessageDTO messageDto = bioDedupeProcessor.process(dto, stageName);
@@ -779,9 +779,9 @@ public class BioDedupeProcessorTest {
 
 		when(abisHandlerUtil.getProcessedMatchedResult(anyString(), anyString(), anyInt(), anyString(), any()))
 				.thenReturn(processedMatchedResult);
-		when(utilities.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
+		when(utility.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
 				.thenReturn(false);
-		when(utilities.allBiometricHaveException(anyString(), anyString(), any()))
+		when(utility.allBiometricHaveException(anyString(), anyString(), any()))
 				.thenReturn(false);
 
 		MessageDTO messageDto = bioDedupeProcessor.process(dto, stageName);
@@ -817,9 +817,9 @@ public class BioDedupeProcessorTest {
 
 		when(abisHandlerUtil.getProcessedMatchedResult(anyString(), anyString(), anyInt(), anyString(), any()))
 				.thenReturn(processedMatchedResult);
-		when(utilities.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
+		when(utility.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
 				.thenReturn(false);
-		when(utilities.allBiometricHaveException(anyString(), anyString(), any()))
+		when(utility.allBiometricHaveException(anyString(), anyString(), any()))
 				.thenReturn(true);
 
 		MessageDTO messageDto = bioDedupeProcessor.process(dto, stageName);
@@ -855,7 +855,7 @@ public class BioDedupeProcessorTest {
 
 		when(abisHandlerUtil.getProcessedMatchedResult(anyString(), anyString(), anyInt(), anyString(), any()))
 				.thenReturn(processedMatchedResult);
-		when(utilities.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
+		when(utility.wasInfantWhenLastPacketProcessed(anyString(), anyString(), any()))
 				.thenReturn(true);
 
 		MessageDTO messageDto = bioDedupeProcessor.process(dto, stageName);
@@ -937,7 +937,7 @@ public class BioDedupeProcessorTest {
 		processedMatchedResult.setBiometricMatchedForPacketUIN(false);
 		Mockito.when(abisHandlerUtil.getProcessedMatchedResult(any(), any(), anyInt(), any(), any())).thenReturn(processedMatchedResult);
 
-		when(utilities.allBiometricHaveException(anyString(), anyString(), any()))
+		when(utility.allBiometricHaveException(anyString(), anyString(), any()))
 				.thenThrow(new BiometricClassificationException("BDD-ERR", "Biometric classification exception"));
 
 		MessageDTO result = bioDedupeProcessor.process(dto, stageName);
