+++ registration-processor/core-processor/registration-processor-uin-generator-stage/src/main/java/io/mosip/registration/processor/stages/uingenerator/stage/UinGeneratorStage.java
@@ -1178,7 +1178,7 @@ public class UinGeneratorStage extends MosipVerticleAPIManager {
 		}
 
 		// Try to fetch the key using getMappedFieldName
-		String packetCreatedOnKey = utilities.getMappedFieldName(MappingJsonConstants.PACKET_CREATED_ON);
+		String packetCreatedOnKey = utility.getMappedFieldName(MappingJsonConstants.PACKET_CREATED_ON);
 
 		if (packetCreatedOnKey == null) {
 			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
@@ -1187,7 +1187,7 @@ public class UinGeneratorStage extends MosipVerticleAPIManager {
 		}
 
 		// Fallback to metaInfo if not present in packet
-		String packetCreatedOn = utilities.retrieveCreatedDateFromPacket(
+		String packetCreatedOn = utility.retrieveCreatedDateFromPacket(
 				registrationId,
 				registrationStatusDto.getRegistrationType(),
 				ProviderStageName.UIN_GENERATOR
